Fixes and Requested Changes to Component Names ... - bug fix for DWA and DWAPDB components ... were previously using 'data.dwaNumber' and 'data.pdbNumber' in their component names, except those fields don't exist - changed to using 'data.typeRecordNumber', which is the correct thing anyway - all populated board kits and the multi-type shipment will now have the 'data.typeRecordNumber' in their names (will need to run the Admin Utility to update existing components) - moved the current Admin Utility function for setting component names and DUNE PIDs to a new library file ... expecting to have more such functions in the future, so best to keep them in the same place (updated route to point to new library file as well)
